+++ tiles/src/core/account/atproto.rs
@@ -1,6 +1,6 @@
 //! Handles atprotocol stuff
 
-use anyhow::{Result, anyhow};
+use anyhow::{Context, Result, anyhow};
 use atrium_api::{
     agent::Agent,
     types::{Unknown, string::Did},
@@ -30,8 +30,10 @@ use std::error::Error;
 use hickory_resolver::TokioResolver;
 
 use crate::{
-    core::storage::db::Dbconn, daemon::start_internal_server, repl::SharedSession,
-    utils::get_unix_time_now,
+    core::storage::db::Dbconn,
+    daemon::start_internal_server,
+    repl::SharedSession,
+    utils::{crypto::encrypt_to_base64, get_unix_time_now},
 };
 
 // TODO: Make this dynamic porting
@@ -304,7 +306,7 @@ pub fn fetch_logged_in_data(conn: &Connection) -> Result<Option<AtprotoAuthData>
         [],
         |row| {
             Ok(AtprotoAuthData {
-                key: row.get(0)?,
+               key: row.get(0)?,
                 session: row.get(1)?,
                 state: row.get(2)?,
                 is_logged_in: row.get(3)?,
@@ -317,7 +319,12 @@ pub fn fetch_logged_in_data(conn: &Connection) -> Result<Option<AtprotoAuthData>
 }
 
 //TODO: Move the login check to common fn
-pub async fn share_session(conn: &Connection, shared_session: SharedSession) -> Result<()> {
+// TODO: Add tests for share session plss
+pub async fn share_session(
+    conn: &Connection,
+    shared_session: SharedSession,
+    is_private: bool,
+) -> Result<()> {
     if let Some(auth_data) = fetch_logged_in_data(conn)? {
         let (client, mem_session_store) = create_oauth_client()?;
         let session: Session = serde_json::from_str(&auth_data.session)?;
@@ -326,15 +333,36 @@ pub async fn share_session(conn: &Connection, shared_session: SharedSession) -> 
 
         mem_session_store.set(did_struct.clone(), session).await?;
 
-        println!("Writing to PDS and generating link...");
-        //TODO: Add a user friendly err latta
-        let oauth_session = client.restore(&did_struct).await?;
+        let write_info = if is_private {
+            "Writing to PDS with encrypted content and generating private link..."
+        } else {
+            "Writing to PDS and generating link..."
+        };
+        println!("{}", write_info);
+        let oauth_session = client
+            .restore(&did_struct)
+            .await
+            .context("Failed to restore the oauth session")?;
+
         let agent = Agent::new(oauth_session);
 
-        let shared_session_value = serde_json::to_value(shared_session)?;
+        let (shared_session_value, nonce, key) = if is_private {
+            let encrypted_content = encrypt_to_base64(&serde_json::to_vec(&shared_session)?)
+                .context("Failed to encrypt the session")?;
+            let ciphertxt = encrypted_content.ciphertext;
+            (
+                serde_json::json!({
+                    "enc_content": ciphertxt
+                }),
+                Some(encrypted_content.nonce),
+                Some(encrypted_content.key),
+            )
+        } else {
+            (serde_json::to_value(shared_session)?, None, None)
+        };
+
         let record: Unknown = serde_json::from_value(shared_session_value)?;
 
-        //TODO: can we remove the unwrap at collection
         let create_result = agent
             .api
             .com
@@ -342,7 +370,9 @@ pub async fn share_session(conn: &Connection, shared_session: SharedSession) -> 
             .repo
             .create_record(
                 atrium_api::com::atproto::repo::create_record::InputData {
-                    collection: "run.tiles.session".parse().unwrap(),
+                    collection: "run.tiles.session"
+                        .parse()
+                        .map_err(|_e| anyhow!("Failed to parse to nsid"))?,
                     repo: did_struct.clone().into(),
                     rkey: None,
                     record,
@@ -357,7 +387,19 @@ pub async fn share_session(conn: &Connection, shared_session: SharedSession) -> 
 
         let base_encoded_at_url = data_encoding::BASE64.encode(url.as_bytes());
 
-        let shareable_url = format!("https://tiles.run/share/{}", base_encoded_at_url);
+        let shareable_base_url = format!("https://tiles.run/share/{}", base_encoded_at_url);
+
+        let shareable_url = if is_private {
+            format!(
+                "{}#{}.{}",
+                shareable_base_url,
+                nonce.expect("Nonce not found"),
+                key.expect("Key not found")
+            )
+        } else {
+            shareable_base_url
+        };
+
         println!("successfully posted at {}", shareable_url);
 
         // Updating the session token