+++ tiles/src/utils/mod.rs
@@ -1,9 +1,9 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
 pub mod config;
+pub mod crypto;
 pub mod hf_model_downloader;
 pub mod installer;
-
 pub fn get_unix_time_now() -> u64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)